Renaming two folders to Events from Activities (#822)
