+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/DeleteEventCommand.cs
