+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/DeleteEventCommandHandler.cs
