+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/EditEventCommand.cs
