+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/EditEventCommandHandler.cs
