+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/EventDetailQuery.cs
