+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/EventDetailQueryHandler.cs
