+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/ManagingOrganizationIdByEventIdQuery.cs
