+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Events/ManagingOrganizationIdByEventIdQueryHandler.cs
