+++ AllReadyApp/Web-App/AllReady/Features/Events/AddEventSignupCommandHandlerAsync.cs
