+++ AllReadyApp/Web-App/AllReady/Features/Events/EventByEventIdQuery.cs
