+++ AllReadyApp/Web-App/AllReady/Features/Events/EventByEventIdQueryHandler.cs
