+++ AllReadyApp/Web-App/AllReady/Features/Events/EventSignupByEventIdAndUserIdQuery.cs
