+++ AllReadyApp/Web-App/AllReady/Features/Events/EventSignupByEventIdAndUserIdQueryHandler.cs
