+++ AllReadyApp/Web-App/AllReady/Features/Events/EventSignupCommand.cs
