+++ AllReadyApp/Web-App/AllReady/Features/Events/EventSignupHandler.cs
