+++ AllReadyApp/Web-App/AllReady/Features/Events/EventsByGeographyQuery.cs
