+++ AllReadyApp/Web-App/AllReady/Features/Events/EventsByGeographyQueryHandler.cs
