+++ AllReadyApp/Web-App/AllReady/Features/Events/GetMyEventsQuery.cs
