+++ AllReadyApp/Web-App/AllReady/Features/Events/GetMyEventsQueryHandler.cs
