+++ AllReadyApp/Web-App/AllReady/Features/Events/ShowEventQuery.cs
