+++ AllReadyApp/Web-App/AllReady/Features/Events/ShowEventQueryHandler.cs
