+++ AllReadyApp/Web-App/AllReady/Features/Events/UnregisterEvent.cs
