+++ AllReadyApp/Web-App/AllReady/Features/Events/UnregisterEventHandler.cs
